Added profile image URL request for Facebook

+++ src/com/chalmers/feedlr/client/TwitterHelper.java
@@ -61,7 +61,7 @@ public class TwitterHelper {
 		return TwitterJSONParser.parseTweets(response);
 	}
 
-	public List<TwitterItem> getUserTweets(int userID) {
+	public List<TwitterItem> getUserTweets(long userID) {
 		long time = System.currentTimeMillis();
 
 		StringBuilder url = new StringBuilder();
